Adding TiffMapForReading / TiffMapForWriting; requiring these classes instead of TiffMap while using TiffReader/TiffWriter

+++ algart-scifio-bridge/src/test/java/net/algart/matrices/tiff/tests/scifio/PureScifioTiffReadWriteTest.java
@@ -80,7 +80,7 @@ public class PureScifioTiffReadWriteTest {
 //                bytes = TiffTools.unpackUnusualPrecisions(bytes, ifd, numberOfChannels, w * h);
                 bytes = TiffMap.toInterleavedBytes(
                         bytes, numberOfChannels, FormatTools.getBytesPerPixel(ifd.getPixelType()),
-                        w * h);
+                        (long) w * (long) h);
                 boolean last = ifdIndex == ifdList.size() - 1;
                 final IFD newIfd = removeUndesirableTags(ifd);
